Admitters: move storage related checks from vm admitter
Move from the vm admitter the checks for the datavolumetemplate
and datavolume to the storage admitter pkg.
Also move snapshot and restore in progress checks
on the vm status.
