+++ pkg/virt-api/webhooks/validating-webhook/admitters/vms-admitter.go
@@ -23,36 +23,30 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"reflect"
-
-	"kubevirt.io/kubevirt/pkg/defaults"
-	"kubevirt.io/kubevirt/pkg/virt-config/deprecation"
-
-	corev1 "k8s.io/api/core/v1"
 
 	admissionv1 "k8s.io/api/admission/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/validation/field"
 	k8sfield "k8s.io/apimachinery/pkg/util/validation/field"
 	"k8s.io/client-go/tools/cache"
 
 	v1 "kubevirt.io/api/core/v1"
 	instancetypev1beta1 "kubevirt.io/api/instancetype/v1beta1"
 	"kubevirt.io/client-go/kubecli"
-	cdiv1 "kubevirt.io/containerized-data-importer-api/pkg/apis/core/v1beta1"
 
 	"kubevirt.io/kubevirt/pkg/controller"
-	netadmitter "kubevirt.io/kubevirt/pkg/network/admitter"
-	migrationutil "kubevirt.io/kubevirt/pkg/util/migrations"
-
+	"kubevirt.io/kubevirt/pkg/defaults"
 	"kubevirt.io/kubevirt/pkg/instancetype"
 	"kubevirt.io/kubevirt/pkg/liveupdate/memory"
 	metrics "kubevirt.io/kubevirt/pkg/monitoring/metrics/virt-api"
+	netadmitter "kubevirt.io/kubevirt/pkg/network/admitter"
+	storageAdmitters "kubevirt.io/kubevirt/pkg/storage/admitters"
+	migrationutil "kubevirt.io/kubevirt/pkg/util/migrations"
 	webhookutils "kubevirt.io/kubevirt/pkg/util/webhooks"
 	"kubevirt.io/kubevirt/pkg/virt-api/webhooks"
 	virtconfig "kubevirt.io/kubevirt/pkg/virt-config"
+	"kubevirt.io/kubevirt/pkg/virt-config/deprecation"
 )
 
 var validRunStrategies = []v1.VirtualMachineRunStrategy{v1.RunStrategyHalted, v1.RunStrategyManual, v1.RunStrategyAlways, v1.RunStrategyRerunOnFailure, v1.RunStrategyOnce}
@@ -145,7 +139,7 @@ func (admitter *VMsAdmitter) Admit(ctx context.Context, ar *admissionv1.Admissio
 		return webhookutils.ToAdmissionResponse(causes)
 	}
 
-	causes, err = admitter.validateVirtualMachineDataVolumeTemplateNamespace(ar.Request, &vm)
+	causes, err = storageAdmitters.Admit(admitter.VirtClient, ctx, ar.Request, &vm, admitter.ClusterConfig)
 	if err != nil {
 		return webhookutils.ToAdmissionResponseError(err)
 	}
@@ -157,16 +151,6 @@ func (admitter *VMsAdmitter) Admit(ctx context.Context, ar *admissionv1.Admissio
 	if err != nil {
 		return webhookutils.ToAdmissionResponseError(err)
 	} else if len(causes) > 0 {
-		return webhookutils.ToAdmissionResponse(causes)
-	}
-
-	causes = validateSnapshotStatus(ar.Request, &vm)
-	if len(causes) > 0 {
-		return webhookutils.ToAdmissionResponse(causes)
-	}
-
-	causes = validateRestoreStatus(ar.Request, &vm)
-	if len(causes) > 0 {
 		return webhookutils.ToAdmissionResponse(causes)
 	}
 
@@ -199,12 +183,7 @@ func (admitter *VMsAdmitter) AdmitStatus(ctx context.Context, ar *admissionv1.Ad
 		return webhookutils.ToAdmissionResponse(causes)
 	}
 
-	causes = validateSnapshotStatus(ar.Request, vm)
-	if len(causes) > 0 {
-		return webhookutils.ToAdmissionResponse(causes)
-	}
-
-	causes = validateRestoreStatus(ar.Request, vm)
+	causes = storageAdmitters.AdmitStatus(admitter.VirtClient, ctx, ar.Request, vm, admitter.ClusterConfig)
 	if len(causes) > 0 {
 		return webhookutils.ToAdmissionResponse(causes)
 	}
@@ -297,42 +276,6 @@ func (admitter *VMsAdmitter) applyInstancetypeToVm(vm *v1.VirtualMachine) (*inst
 	return nil, nil, causes
 }
 
-func (admitter *VMsAdmitter) validateVirtualMachineDataVolumeTemplateNamespace(ar *admissionv1.AdmissionRequest, vm *v1.VirtualMachine) ([]metav1.StatusCause, error) {
-	var causes []metav1.StatusCause
-
-	if ar.Operation == admissionv1.Update || ar.Operation == admissionv1.Delete {
-		oldVM := &v1.VirtualMachine{}
-		if err := json.Unmarshal(ar.OldObject.Raw, oldVM); err != nil {
-			return []metav1.StatusCause{{
-				Type:    metav1.CauseTypeUnexpectedServerResponse,
-				Message: "Could not fetch old VM",
-			}}, nil
-		}
-
-		if equality.Semantic.DeepEqual(oldVM.Spec.DataVolumeTemplates, vm.Spec.DataVolumeTemplates) {
-			return nil, nil
-		}
-	}
-
-	for idx, dataVolume := range vm.Spec.DataVolumeTemplates {
-		targetNamespace := vm.Namespace
-		if targetNamespace == "" {
-			targetNamespace = ar.Namespace
-		}
-		if dataVolume.Namespace != "" && dataVolume.Namespace != targetNamespace {
-			causes = append(causes, metav1.StatusCause{
-				Type:    metav1.CauseTypeFieldValueInvalid,
-				Message: fmt.Sprintf("Embedded DataVolume namespace %s differs from VM namespace %s", dataVolume.Namespace, targetNamespace),
-				Field:   k8sfield.NewPath("spec", "dataVolumeTemplates").Index(idx).String(),
-			})
-
-			continue
-		}
-	}
-
-	return causes, nil
-}
-
 func ValidateVirtualMachineSpec(field *k8sfield.Path, spec *v1.VirtualMachineSpec, config *virtconfig.ClusterConfig, isKubeVirtServiceAccount bool) []metav1.StatusCause {
 	var causes []metav1.StatusCause
 
@@ -347,123 +290,11 @@ func ValidateVirtualMachineSpec(field *k8sfield.Path, spec *v1.VirtualMachineSpe
 	causes = append(causes, ValidateVirtualMachineInstanceMetadata(field.Child("template", "metadata"), &spec.Template.ObjectMeta, config, isKubeVirtServiceAccount)...)
 	causes = append(causes, ValidateVirtualMachineInstanceSpec(field.Child("template", "spec"), &spec.Template.Spec, config)...)
 
-	causes = append(causes, validateDataVolumeTemplate(field, spec)...)
+	causes = append(causes, storageAdmitters.ValidateDataVolumeTemplate(field, spec)...)
 	causes = append(causes, validateRunStrategy(field, spec)...)
 	causes = append(causes, validateLiveUpdateFeatures(field, spec, config)...)
 
 	return causes
-}
-
-func validateDataVolumeTemplate(field *k8sfield.Path, spec *v1.VirtualMachineSpec) (causes []metav1.StatusCause) {
-	for idx, dataVolume := range spec.DataVolumeTemplates {
-		cause := validateDataVolume(field.Child("dataVolumeTemplate").Index(idx), dataVolume)
-		if cause != nil {
-			causes = append(causes, cause...)
-			continue
-		}
-
-		dataVolumeRefFound := false
-		for _, volume := range spec.Template.Spec.Volumes {
-			if volume.VolumeSource.PersistentVolumeClaim != nil && volume.VolumeSource.PersistentVolumeClaim.ClaimName == dataVolume.Name ||
-				volume.VolumeSource.DataVolume != nil && volume.VolumeSource.DataVolume.Name == dataVolume.Name {
-				dataVolumeRefFound = true
-				break
-			}
-		}
-
-		if !dataVolumeRefFound {
-			causes = append(causes, metav1.StatusCause{
-				Type:    metav1.CauseTypeFieldValueRequired,
-				Message: fmt.Sprintf("DataVolumeTemplate entry %s must be referenced in the VMI template's 'volumes' list", field.Child("dataVolumeTemplate").Index(idx).String()),
-				Field:   field.Child("dataVolumeTemplate").Index(idx).String(),
-			})
-		}
-	}
-	return causes
-}
-
-func validateDataVolume(field *k8sfield.Path, dataVolume v1.DataVolumeTemplateSpec) []metav1.StatusCause {
-	if dataVolume.Name == "" {
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeFieldValueRequired,
-			Message: fmt.Sprintf("'name' field must not be empty for DataVolumeTemplate entry %s.", field.Child("name").String()),
-			Field:   field.Child("name").String(),
-		}}
-	}
-	if dataVolume.Spec.PVC == nil && dataVolume.Spec.Storage == nil {
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeFieldValueInvalid,
-			Message: "Missing Data volume PVC or Storage",
-			Field:   field.Child("PVC", "Storage").String(),
-		}}
-	}
-	if dataVolume.Spec.PVC != nil && dataVolume.Spec.Storage != nil {
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeFieldValueInvalid,
-			Message: "Duplicate storage definition, both target storage and target pvc defined",
-			Field:   field.Child("PVC", "Storage").String(),
-		}}
-	}
-
-	var dataSourceRef *corev1.TypedObjectReference
-	var dataSource *corev1.TypedLocalObjectReference
-	if dataVolume.Spec.PVC != nil {
-		dataSourceRef = dataVolume.Spec.PVC.DataSourceRef
-		dataSource = dataVolume.Spec.PVC.DataSource
-	} else if dataVolume.Spec.Storage != nil {
-		dataSourceRef = dataVolume.Spec.Storage.DataSourceRef
-		dataSource = dataVolume.Spec.Storage.DataSource
-	}
-
-	// dataVolume is externally populated
-	if (dataSourceRef != nil || dataSource != nil) &&
-		(dataVolume.Spec.Source != nil || dataVolume.Spec.SourceRef != nil) {
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeFieldValueInvalid,
-			Message: "External population is incompatible with Source and SourceRef",
-			Field:   field.Child("source").String(),
-		}}
-	}
-
-	if (dataSourceRef == nil && dataSource == nil) &&
-		(dataVolume.Spec.Source == nil && dataVolume.Spec.SourceRef == nil) {
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeFieldValueInvalid,
-			Message: "Data volume should have either Source, SourceRef, or be externally populated",
-			Field:   field.Child("source", "sourceRef").String(),
-		}}
-	}
-
-	if dataVolume.Spec.Source != nil {
-		return validateNumberOfSources(field, dataVolume.Spec.Source)
-	}
-
-	return nil
-}
-
-func validateNumberOfSources(field *field.Path, source *cdiv1.DataVolumeSource) []metav1.StatusCause {
-	numberOfSources := 0
-	s := reflect.ValueOf(source).Elem()
-	for i := 0; i < s.NumField(); i++ {
-		if !reflect.ValueOf(s.Field(i).Interface()).IsNil() {
-			numberOfSources++
-		}
-	}
-	if numberOfSources == 0 {
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeFieldValueInvalid,
-			Message: "Missing dataVolume valid source",
-			Field:   field.Child("source").String(),
-		}}
-	}
-	if numberOfSources > 1 {
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeFieldValueInvalid,
-			Message: "Multiple dataVolume sources",
-			Field:   field.Child("source").String(),
-		}}
-	}
-	return nil
 }
 
 func validateRunStrategy(field *k8sfield.Path, spec *v1.VirtualMachineSpec) (causes []metav1.StatusCause) {
@@ -735,92 +566,4 @@ func validateDiskConfiguration(disk *v1.Disk, name string) []metav1.StatusCause 
 	}
 
 	return nil
-}
-
-func validateRestoreStatus(ar *admissionv1.AdmissionRequest, vm *v1.VirtualMachine) []metav1.StatusCause {
-	if ar.Operation != admissionv1.Update || vm.Status.RestoreInProgress == nil {
-		return nil
-	}
-
-	oldVM := &v1.VirtualMachine{}
-	if err := json.Unmarshal(ar.OldObject.Raw, oldVM); err != nil {
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeUnexpectedServerResponse,
-			Message: "Could not fetch old VM",
-		}}
-	}
-
-	if !equality.Semantic.DeepEqual(oldVM.Spec, vm.Spec) {
-		strategy, _ := vm.RunStrategy()
-		if strategy != v1.RunStrategyHalted {
-			return []metav1.StatusCause{{
-				Type:    metav1.CauseTypeFieldValueNotSupported,
-				Message: fmt.Sprintf("Cannot start VM until restore %q completes", *vm.Status.RestoreInProgress),
-				Field:   k8sfield.NewPath("spec").String(),
-			}}
-		}
-	}
-
-	return nil
-}
-
-func validateSnapshotStatus(ar *admissionv1.AdmissionRequest, vm *v1.VirtualMachine) []metav1.StatusCause {
-	if ar.Operation != admissionv1.Update || vm.Status.SnapshotInProgress == nil {
-		return nil
-	}
-
-	oldVM := &v1.VirtualMachine{}
-	if err := json.Unmarshal(ar.OldObject.Raw, oldVM); err != nil {
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeUnexpectedServerResponse,
-			Message: "Could not fetch old VM",
-		}}
-	}
-
-	if !compareVolumes(oldVM.Spec.Template.Spec.Volumes, vm.Spec.Template.Spec.Volumes) {
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeFieldValueNotSupported,
-			Message: fmt.Sprintf("Cannot update VM disks or volumes until snapshot %q completes", *vm.Status.SnapshotInProgress),
-			Field:   k8sfield.NewPath("spec").String(),
-		}}
-	}
-	if !compareRunningSpec(&oldVM.Spec, &vm.Spec) {
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeFieldValueNotSupported,
-			Message: fmt.Sprintf("Cannot update VM running state until snapshot %q completes", *vm.Status.SnapshotInProgress),
-			Field:   k8sfield.NewPath("spec").String(),
-		}}
-	}
-
-	return nil
-}
-
-func compareVolumes(old, new []v1.Volume) bool {
-	if len(old) != len(new) {
-		return false
-	}
-
-	for i, volume := range old {
-		if !equality.Semantic.DeepEqual(volume, new[i]) {
-			return false
-		}
-	}
-
-	return true
-}
-
-func compareRunningSpec(old, new *v1.VirtualMachineSpec) bool {
-	if old == nil || new == nil {
-		// This should never happen, but just in case return false
-		return false
-	}
-
-	// Its impossible to get here while both running and RunStrategy are nil.
-	if old.Running != nil && new.Running != nil {
-		return *old.Running == *new.Running
-	}
-	if old.RunStrategy != nil && new.RunStrategy != nil {
-		return *old.RunStrategy == *new.RunStrategy
-	}
-	return false
 }
